+++ NEWS.md
@@ -1,3 +1,6 @@
-# octopus 0.1.2
+# octopus 0.2.1
 
-* Added a `NEWS.md` file to track changes to the package.
+* Added support for Microsoft SQL Server Database
+* Added support for hot key commands to run queries (ctr+shift+enter)
+* Minor bug fixes
+